fix: DB에 유저 선호 음식 저장하는 로직 추가

+++ src/main/java/umc/domain/member/dto/MemberReqDTO.java
@@ -3,7 +3,6 @@ package umc.domain.member.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Past;
-import jakarta.validation.constraints.Pattern;
 import umc.domain.member.enums.Gender;
 
 import java.time.LocalDate;
@@ -27,7 +26,7 @@ public class MemberReqDTO {
             LocalDate birth,
             @NotBlank
             String address,
-            List<Integer> userFood,
+            List<Long> userFood,
             @NotBlank
             String email,
             @NotBlank
